fix: Reflection flicker.
half4 in RootConstants do not work correctly.
Swapped to float4 to fix the issue.

+++ Sample/ReflectionProbe.hpp
@@ -23,10 +23,10 @@ class ReflectionProbe
 public:
   struct Probe
   {
-    Float16 PositionX;
-    Float16 PositionY;
-    Float16 PositionZ;
-    Float16 CaptureRadius;
+    float PositionX;
+    float PositionY;
+    float PositionZ;
+    float CaptureRadius;
   };
 
 private:
